CohortMethodData-class {CohortMethod}	R Documentation

Cohort Method Data

Description

CohortMethodData is an S4 class that inherits from CoviarateData, which in turn inherits from Andromeda. It contains information on the cohorts, their outcomes, and baseline covariates. Information about multiple outcomes can be captured at once for efficiency reasons.

A CohortMethodData is typically created using getDbCohortMethodData(), can only be saved using saveCohortMethodData(), and loaded using loadCohortMethodData().

Usage

## S4 method for signature 'CohortMethodData'
show(object)

## S4 method for signature 'CohortMethodData'
summary(object)

Arguments

object
An object of type CohortMethodData.

[Package CohortMethod version 4.0.0 Index]