computePsAuc {CohortMethod}	R Documentation

Compute the area under the ROC curve

Description

Compute the area under the ROC curve of the propensity score.

Usage

computePsAuc(data, confidenceIntervals = FALSE)

Arguments

data
A data frame with at least the two columns described below

confidenceIntervals
Compute 95 percent confidence intervals (computationally expensive for large data sets)

Details

The data frame should have a least the following two columns:

treatment (integer): Column indicating whether the person is in the target (1) or comparator (0) group.

propensityScore (numeric): Propensity score.

Value

A tibble holding the AUC and its 95 percent confidence interval

Examples

treatment <- rep(0:1, each = 100)
propensityScore <- c(rnorm(100, mean = 0.4, sd = 0.25), rnorm(100, mean = 0.6, sd = 0.25))
data <- data.frame(treatment = treatment, propensityScore = propensityScore)
data <- data[data$propensityScore > 0 & data$propensityScore < 1, ]
computePsAuc(data)

[Package CohortMethod version 4.0.0 Index]