Create a CohortMethod analysis specification

Create a CohortMethod analysis specification

createCmAnalysis(analysisId = 1, description = "", targetType = NULL,
  comparatorType = NULL, getDbCohortMethodDataArgs, createStudyPopArgs,
  createPs = FALSE, createPsArgs = NULL, trimByPs = FALSE,
  trimByPsArgs = NULL, trimByPsToEquipoise = FALSE,
  trimByPsToEquipoiseArgs = NULL, matchOnPs = FALSE,
  matchOnPsArgs = NULL, matchOnPsAndCovariates = FALSE,
  matchOnPsAndCovariatesArgs = NULL, stratifyByPs = FALSE,
  stratifyByPsArgs = NULL, stratifyByPsAndCovariates = FALSE,
  stratifyByPsAndCovariatesArgs = NULL, fitOutcomeModel = FALSE,
  fitOutcomeModelArgs = NULL)

Arguments

analysisId
An integer that will be used later to refer to this specific set of analysis choices.

description
A short description of the analysis.

targetType
If more than one target is provided for each drugComparatorOutcome, this field should be used to select the specific target to use in this analysis.

comparatorType
If more than one comparator is provided for each drugComparatorOutcome, this field should be used to select the specific comparator to use in this analysis.

getDbCohortMethodDataArgs
An object representing the arguments to be used when calling the getDbCohortMethodData function.

createStudyPopArgs
An object representing the arguments to be used when calling the createStudyPopulation function.

createPs
Should the createPs function be used in this analysis?

createPsArgs
An object representing the arguments to be used when calling the createPs function.

trimByPs
Should the trimByPs function be used in this analysis?

trimByPsArgs
An object representing the arguments to be used when calling the trimByPs function.

trimByPsToEquipoise
Should the trimByPsToEquipoise function be used in this analysis?

trimByPsToEquipoiseArgs
An object representing the arguments to be used when calling the trimByPsToEquipoise function.

matchOnPs
Should the matchOnPs function be used in this analysis?

matchOnPsArgs
An object representing the arguments to be used when calling the matchOnPs function.

matchOnPsAndCovariates
Should the matchOnPsAndCovariates function be used in this analysis?

matchOnPsAndCovariatesArgs
An object representing the arguments to be used when calling the matchOnPsAndCovariates function.

stratifyByPs
Should the stratifyByPs function be used in this analysis?

stratifyByPsArgs
An object representing the arguments to be used when calling the stratifyByPs function.

stratifyByPsAndCovariates
Should the stratifyByPsAndCovariates function be used in this analysis?

stratifyByPsAndCovariatesArgs
An object representing the arguments to be used when calling the stratifyByPsAndCovariates function.

fitOutcomeModel
Should the fitOutcomeModel function be used in this analysis?

fitOutcomeModelArgs
An object representing the arguments to be used when calling the fitOutcomeModel function.

Details

Create a set of analysis choices, to be used with the runCmAnalyses function.