Create CohortMethod diagnostics thresholds

Threshold used when calling exportToCsv() to determine if we pass or fail diagnostics.

createCmDiagnosticThresholds(
  mdrrThreshold = 10,
  easeThreshold = 0.25,
  sdmThreshold = 0.1,
  equipoiseThreshold = 0.2,
  attritionFractionThreshold = NULL,
  generalizabilitySdmThreshold = 1
)

Arguments

mdrrThreshold

What is the maximum allowed minimum detectable relative risk (MDRR)?

easeThreshold

What is the maximum allowed expected absolute systematic error (EASE).

sdmThreshold

What is the maximum allowed standardized difference of mean (SDM)? If any covariate has an SDM exceeding this threshold, the diagnostic will fail.

equipoiseThreshold

What is the minimum required equipoise?

attritionFractionThreshold

DEPRECATED. See generalizabilitySdmThreshold instead.

generalizabilitySdmThreshold

What is the maximum allowed standardized difference of mean (SDM)when comparing the population before and after PS adjustments? If the SDM is greater than this value, the diagnostic will fail.

Value

An object of type CmDiagnosticThresholds.