Create a parameter object for the function createStudyPopulation

Create a parameter object for the function createStudyPopulation

createCreateStudyPopulationArgs(
  firstExposureOnly = FALSE,
  restrictToCommonPeriod = FALSE,
  washoutPeriod = 0,
  removeDuplicateSubjects = FALSE,
  removeSubjectsWithPriorOutcome = TRUE,
  priorOutcomeLookback = 99999,
  minDaysAtRisk = 1,
  maxDaysAtRisk = 99999,
  riskWindowStart = 0,
  addExposureDaysToStart = NULL,
  startAnchor = "cohort start",
  riskWindowEnd = 0,
  addExposureDaysToEnd = NULL,
  endAnchor = "cohort end",
  censorAtNewRiskWindow = FALSE
)

Arguments

firstExposureOnly
Should only the first exposure per subject be included?

restrictToCommonPeriod
Restrict the analysis to the period when both exposures are observed?

washoutPeriod
The minimum required continuous observation time prior to index date for a person to be included in the cohort.

removeDuplicateSubjects
Remove subjects that are in both the target and comparator cohort? See details for allowed values.

removeSubjectsWithPriorOutcome
Remove subjects that have the outcome prior to the risk window start?

priorOutcomeLookback
How many days should we look back when identifying prior outcomes?

minDaysAtRisk
The minimum required number of days at risk. Risk windows with fewer days than this number are removed from the analysis.

maxDaysAtRisk
The maximum allowed number of days at risk. Risk windows that are longer will be truncated to this number of days.

riskWindowStart
The start of the risk window (in days) relative to the startAnchor.

addExposureDaysToStart
DEPRECATED: Add the length of exposure the start of the risk window? Use startAnchor instead.

startAnchor
The anchor point for the start of the risk window. Can be "cohort start" or "cohort end".

riskWindowEnd
The end of the risk window (in days) relative to the endAnchor.

addExposureDaysToEnd
DEPRECATED: Add the length of exposure the risk window? Use endAnchor instead.

endAnchor
The anchor point for the end of the risk window. Can be "cohort start" or "cohort end".

censorAtNewRiskWindow
If a subject is in multiple cohorts, should time-at-risk be censored when the new time-at-risk starts to prevent overlap?

Details

Create an object defining the parameter values.