computeCovariateBalance {CohortMethod}	R Documentation

Compute covariate balance before and after matching and trimming

Description

For every covariate, prevalence in treatment and comparator groups before and after matching/trimming are computed. When variable ratio matching was used the balance score will be corrected according the method described in Austin et al (2008).

Usage

computeCovariateBalance(
  population,
  cohortMethodData,
  subgroupCovariateId = NULL
)

Arguments

population
A data frame containing the people that are remaining after matching and/or trimming.

cohortMethodData
An object of type CohortMethodData as generated using getDbCohortMethodData().

subgroupCovariateId
Optional: a covariate ID of a binary covariate that indicates a subgroup of interest. Both the before and after populations will be restricted to this subgroup before computing covariate balance.

Details

The population data frame should have the following three columns:

rowId (numeric): A unique identifier for each row (e.g. the person ID).

treatment (integer): Column indicating whether the person is in the target (1) or comparator (0) group.

propensityScore (numeric): Propensity score.

Value

Returns a tibble describing the covariate balance before and after matching/trimming.

References

Austin, P.C. (2008) Assessing balance in measured baseline covariates when using many-to-one matching on the propensity-score. Pharmacoepidemiology and Drug Safety, 17: 1218-1225.

[Package CohortMethod version 4.0.0 Index]