Create a CohortMethod analysis specification

Create a CohortMethod analysis specification

Usage

createCmAnalysis(
  analysisId = 1,
  description = "",
  getDbCohortMethodDataArgs,
  createStudyPopArgs,
  createPsArgs = NULL,
  trimByPsArgs = NULL,
  trimByPsToEquipoiseArgs = NULL,
  trimByIptwArgs = NULL,
  truncateIptwArgs = NULL,
  matchOnPsArgs = NULL,
  matchOnPsAndCovariatesArgs = NULL,
  stratifyByPsArgs = NULL,
  stratifyByPsAndCovariatesArgs = NULL,
  computeSharedCovariateBalanceArgs = NULL,
  computeCovariateBalanceArgs = NULL,
  fitOutcomeModelArgs = NULL
)

Arguments

analysisId

An integer that will be used later to refer to this specific set of analysis choices.

description

A short description of the analysis.

getDbCohortMethodDataArgs

An object representing the arguments to be used when calling the getDbCohortMethodData() function.

createStudyPopArgs

An object representing the arguments to be used when calling the createStudyPopulation() function.

createPsArgs

An object representing the arguments to be used when calling the createPs() function.

trimByPsArgs

An object representing the arguments to be used when calling the trimByPs() function.

trimByPsToEquipoiseArgs

An object representing the arguments to be used when calling the trimByPsToEquipoise() function.

trimByIptwArgs

An object representing the arguments to be used when calling the trimByIptw() function.

truncateIptwArgs

An object representing the arguments to be used when calling the truncateIptw() function.

matchOnPsArgs

An object representing the arguments to be used when calling the matchOnPs() function.

matchOnPsAndCovariatesArgs

An object representing the arguments to be used when calling the matchOnPsAndCovariates() function.

stratifyByPsArgs

An object representing the arguments to be used when calling the stratifyByPs() function.

stratifyByPsAndCovariatesArgs

An object representing the arguments to be used when calling the stratifyByPsAndCovariates() function.

computeSharedCovariateBalanceArgs

An object representing the arguments to be used when calling the computeCovariateBalance() function per target-comparator-analysis.

computeCovariateBalanceArgs

An object representing the arguments to be used when calling the computeCovariateBalance() function per target-comparator-outcome-analysis.

fitOutcomeModelArgs

An object representing the arguments to be used when calling the fitOutcomeModel() function.

Details

Create a set of analysis choices, to be used with the runCmAnalyses() function.

Providing a NULL value for any of the argument applies the corresponding step will not be executed. For example, if createPsArgs = NULL, no propensity scores will be computed.