createCmDiagnosticThresholds {CohortMethod}	R Documentation

Create CohortMethod diagnostics thresholds

Description

Threshold used when calling exportToCsv() to determine if we pass or fail diagnostics.

Usage

createCmDiagnosticThresholds(
  mdrrThreshold = 10,
  easeThreshold = 0.25,
  sdmThreshold = 0.1,
  equipoiseThreshold = 0.2,
  attritionFractionThreshold = 1
)

Arguments

mdrrThreshold
What is the maximum allowed minimum detectable relative risk (MDRR)?

easeThreshold
What is the maximum allowed expected absolute systematic error (EASE).

sdmThreshold
What is the maximum allowed standardized difference of mean (SDM)? If any covariate has an SDM exceeding this threshold, the diagnostic will fail.

equipoiseThreshold
What is the minimum required equipoise?

attritionFractionThreshold
What is the maximum allowed attrition fraction? If the attrition between the input target cohort and the target cohort entering the outcome model is greater than this fraction, the diagnostic will fail.

Value

An object of type CmDiagnosticThresholds.

[Package CohortMethod version 5.0.0 Index]