createCohortMethodDataSimulationProfile {CohortMethod}	R Documentation

Create simulation profile

Description

Creates a profile based on the provided CohortMethodData object, which can be used to generate simulated data that has similar characteristics.

Usage

createCohortMethodDataSimulationProfile(cohortMethodData)

Arguments

cohortMethodData
An object of type CohortMethodData as generated using getDbCohortMethodData().

Details

The output of this function is an object that can be used by the simulateCohortMethodData() function to generate a cohortMethodData object.

Value

An object of type CohortDataSimulationProfile.

[Package CohortMethod version 4.2.1 Index]