createComputeCovariateBalanceArgs {CohortMethod}	R Documentation

Create a parameter object for the function computeCovariateBalance

Description

Create a parameter object for the function computeCovariateBalance

Usage

createComputeCovariateBalanceArgs(
  subgroupCovariateId = NULL,
  maxCohortSize = 250000,
  covariateFilter = NULL
)

Arguments

subgroupCovariateId
Optional: a covariate ID of a binary covariate that indicates a subgroup of interest. Both the before and after populations will be restricted to this subgroup before computing covariate balance.

maxCohortSize
If the target or comparator cohort are larger than this number, they will be downsampled before computing covariate balance to save time. Setting this number to 0 means no downsampling will be applied.

covariateFilter
Determines the covariates for which to compute covariate balance. Either a vector of covariate IDs, or a table 1 specifications object as generated for example using FeatureExtraction::getDefaultTable1Specifications(). If covariateFilter = NULL, balance will be computed for all variables found in the data.

Details

Create an object defining the parameter values.

[Package CohortMethod version 5.0.0 Index]