createFitOutcomeModelArgs {CohortMethod}	R Documentation

Create a parameter object for the function fitOutcomeModel

Description

Create a parameter object for the function fitOutcomeModel

Usage

createFitOutcomeModelArgs(
  modelType = "logistic",
  stratified = FALSE,
  useCovariates = FALSE,
  inversePtWeighting = FALSE,
  interactionCovariateIds = c(),
  excludeCovariateIds = c(),
  includeCovariateIds = c(),
  prior = createPrior("laplace", useCrossValidation = TRUE),
  control = createControl(cvType = "auto", seed = 1, startingVariance = 0.01, tolerance
    = 2e-07, cvRepetitions = 10, noiseLevel = "quiet")
)

Arguments

modelType
The type of outcome model that will be used. Possible values are "logistic", "poisson", or "cox".

stratified
Should the regression be conditioned on the strata defined in the population object (e.g. by matching or stratifying on propensity scores)?

useCovariates
Whether to use the covariates in the cohortMethodData object in the outcome model.

inversePtWeighting
Use inverse probability of treatment weighting (IPTW)? See details.

interactionCovariateIds
An optional vector of covariate IDs to use to estimate interactions with the main treatment effect.

excludeCovariateIds
Exclude these covariates from the outcome model.

includeCovariateIds
Include only these covariates in the outcome model.

prior
The prior used to fit the model. See Cyclops::createPrior() for details.

control
The control object used to control the cross-validation used to determine the hyperparameters of the prior (if applicable). See Cyclops::createControl() for details.

Details

Create an object defining the parameter values.

[Package CohortMethod version 4.0.0 Index]