createStratifyByPsArgs {CohortMethod}	R Documentation

Create a parameter object for the function stratifyByPs

Description

Create a parameter object for the function stratifyByPs

Usage

createStratifyByPsArgs(
  numberOfStrata = 5,
  stratificationColumns = c(),
  baseSelection = "all"
)

Arguments

numberOfStrata
How many strata? The boundaries of the strata are automatically defined to contain equal numbers of target persons.

stratificationColumns
Names of one or more columns in the data data.frame on which subjects should also be stratified in addition to stratification on propensity score.

baseSelection
What is the base selection of subjects where the strata bounds are to be determined? Strata are defined as equally-sized strata inside this selection. Possible values are "all", "target", and "comparator".

Details

Create an object defining the parameter values.

[Package CohortMethod version 4.0.0 Index]