createStudyPopulation {CohortMethod}	R Documentation

Create a study population

Description

Create a study population

Usage

createStudyPopulation(
  cohortMethodData,
  population = NULL,
  outcomeId,
  firstExposureOnly = FALSE,
  restrictToCommonPeriod = FALSE,
  washoutPeriod = 0,
  removeDuplicateSubjects = FALSE,
  removeSubjectsWithPriorOutcome = TRUE,
  priorOutcomeLookback = 99999,
  minDaysAtRisk = 1,
  riskWindowStart = 0,
  addExposureDaysToStart = NULL,
  startAnchor = "cohort start",
  riskWindowEnd = 0,
  addExposureDaysToEnd = NULL,
  endAnchor = "cohort end",
  censorAtNewRiskWindow = FALSE
)

Arguments

cohortMethodData
An object of type CohortMethodData as generated using getDbCohortMethodData().

population
If specified, this population will be used as the starting point instead of the cohorts in the cohortMethodData object.

outcomeId
The ID of the outcome. If not specified, no outcome-specific transformations will be performed.

firstExposureOnly
Should only the first exposure per subject be included?

restrictToCommonPeriod
Restrict the analysis to the period when both exposures are observed?

washoutPeriod
The minimum required continuous observation time prior to index date for a person to be included in the cohort.

removeDuplicateSubjects
Remove subjects that are in both the target and comparator cohort? See details for allowed values.

removeSubjectsWithPriorOutcome
Remove subjects that have the outcome prior to the risk window start?

priorOutcomeLookback
How many days should we look back when identifying prior outcomes?

minDaysAtRisk
The minimum required number of days at risk.

riskWindowStart
The start of the risk window (in days) relative to the startAnchor.

addExposureDaysToStart
DEPRECATED: Add the length of exposure the start of the risk window? Use startAnchor instead.

startAnchor
The anchor point for the start of the risk window. Can be "cohort start" or "cohort end".

riskWindowEnd
The end of the risk window (in days) relative to the endAnchor.

addExposureDaysToEnd
DEPRECATED: Add the length of exposure the risk window? Use endAnchor instead.

endAnchor
The anchor point for the end of the risk window. Can be "cohort start" or "cohort end".

censorAtNewRiskWindow
If a subject is in multiple cohorts, should time-at-risk be censored when the new time-at-risk starts to prevent overlap?

Details

Create a study population by enforcing certain inclusion and exclusion criteria, defining a risk window, and determining which outcomes fall inside the risk window.

The removeduplicateSubjects argument can have one of the following values:

"keep all": Do not remove subjects that appear in both target and comparator cohort

"keep first": When a subjects appear in both target and comparator cohort, only keep whichever cohort is first in time. If both cohorts start simultaneous, the person is removed from the analysis.

"remove all": Remove subjects that appear in both target and comparator cohort completely from the analysis."

Value

A tibble specifying the study population. This tibble will have the following columns:

rowId: A unique identifier for an exposure.

subjectId: The person ID of the subject.

cohortStartdate: The index date.

outcomeCount The number of outcomes observed during the risk window.

timeAtRisk: The number of days in the risk window.

survivalTime: The number of days until either the outcome or the end of the risk window.

[Package CohortMethod version 4.0.0 Index]