createTargetComparatorOutcomes {CohortMethod}	R Documentation

Create target-comparator-outcomes combinations.

Description

Create target-comparator-outcomes combinations.

Usage

createTargetComparatorOutcomes(
  targetId,
  comparatorId,
  outcomes,
  excludedCovariateConceptIds = c(),
  includedCovariateConceptIds = c()
)

Arguments

targetId
A cohort ID identifying the target exposure in the exposure table.

comparatorId
A cohort ID identifying the comparator exposure in the exposure table.

outcomes
A list of object of type outcome as created by createOutcome().

excludedCovariateConceptIds
A list of concept IDs that cannot be used to construct covariates. This argument is to be used only for exclusion concepts that are specific to the target-comparator combination.

includedCovariateConceptIds
A list of concept IDs that must be used to construct covariates. This argument is to be used only for inclusion concepts that are specific to the target-comparator combination.

Details

Create a set of hypotheses of interest, to be used with the runCmAnalyses() function.

Value

An object of type targetComparatorOutcomes.

[Package CohortMethod version 5.0.0 Index]