createTargetComparatorOutcomes {CohortMethod}	R Documentation

Create target-comparator-outcomes combinations.

Description

Create target-comparator-outcomes combinations.

Usage

createTargetComparatorOutcomes(
  targetId,
  comparatorId,
  outcomeIds,
  excludedCovariateConceptIds = c(),
  includedCovariateConceptIds = c()
)

Arguments

targetId
A concept ID identifying the target drug in the exposure table. If multiple strategies for picking the target will be tested in the analysis, a named list of numbers can be provided instead. In the analysis, the name of the number to be used can be specified using the #' targetType parameter in the createCmAnalysis function.

comparatorId
A concept ID indentifying the comparator drug in the exposure table. If multiple strategies for picking the comparator will be tested in the analysis, a named list of numbers can be provided instead. In the analysis, the name of the number to be used can be specified using the #' comparatorType parameter in the createCmAnalysis function.

outcomeIds
A vector of concept IDs indentifying the outcome(s) in the outcome table.

excludedCovariateConceptIds
A list of concept IDs that cannot be used to construct covariates. This argument is to be used only for exclusion concepts that are specific to the drug-comparator combination.

includedCovariateConceptIds
A list of concept IDs that must be used to construct covariates. This argument is to be used only for inclusion concepts that are specific to the drug-comparator combination.

Details

Create a set of hypotheses of interest, to be used with the runCmAnalyses function.

[Package CohortMethod version 4.0.0 Index]