createTrimByPsArgs {CohortMethod}	R Documentation

Create a parameter object for the function trimByPs

Description

Create a parameter object for the function trimByPs

Usage

createTrimByPsArgs(trimFraction = 0.05)

Arguments

trimFraction
This fraction will be removed from each treatment group. In the target group, persons with the highest propensity scores will be removed, in the comparator group person with the lowest scores will be removed.

Details

Create an object defining the parameter values.

[Package CohortMethod version 4.0.0 Index]