Create an outcome model, and compute the relative risk

fitOutcomeModel creates an outcome model, and computes the relative risk

fitOutcomeModel(population, cohortMethodData = NULL,
  modelType = "logistic", stratified = FALSE, useCovariates = FALSE,
  inversePtWeighting = FALSE, interactionCovariateIds = c(),
  excludeCovariateIds = c(), includeCovariateIds = c(),
  prior = createPrior("laplace", useCrossValidation = TRUE),
  control = createControl(cvType = "auto", seed = 1, startingVariance =
  0.01, tolerance = 2e-07, cvRepetitions = 10, noiseLevel = "quiet"))

Arguments

population
A population object generated by createStudyPopulation, potentially filtered by other functions.

cohortMethodData
An object of type cohortMethodData as generated using getDbCohortMethodData.

modelType
The type of outcome model that will be used. Possible values are "logistic", "poisson", or "cox".

stratified
Should the regression be conditioned on the strata defined in the population object (e.g. by matching or stratifying on propensity scores)?

useCovariates
Whether to use the covariate matrix in the cohortMethodData object in the outcome model.

inversePtWeighting
Use inverse probability of treatment weigting?

interactionCovariateIds
An optional vector of covariate IDs to use to estimate interactions with the main treatment effect.

excludeCovariateIds
Exclude these covariates from the outcome model.

includeCovariateIds
Include only these covariates in the outcome model.

prior
The prior used to fit the model. See createPrior for details.

control
The control object used to control the cross-validation used to determine the hyperparameters of the prior (if applicable). See createControl for details.

Value

An object of class outcomeModel. Generic function summary, coef, and confint are available.