getAttritionTable {CohortMethod}	R Documentation

Get the attrition table for a population

Description

Get the attrition table for a population

Usage

getAttritionTable(object)

Arguments

object
Either an object of type CohortMethodData, a population object generated by functions like createStudyPopulation(), or an object of type outcomeModel.

Value

A tibble specifying the number of people and exposures in the population after specific steps of filtering.

[Package CohortMethod version 4.0.0 Index]